+++ RELEASE_NOTES.md
@@ -1,7 +1,7 @@
 
 # Next Release
 
-- [#135](https://github.com/iiasa/ixmp/pull/135): Test `retixmp` using the R `testthat` package.
+- [#135](https://github.com/iiasa/ixmp/pull/135): Test `rixmp` (former `retixmp`) using the R `testthat` package.
 - [#115](https://github.com/iiasa/ixmp/pull/115): Support iterating with user-supplied callbacks.
 - [#130](https://github.com/iiasa/ixmp/pull/130): Recognize `IXMP_DATA` environment variable for configuration and local databases.
 - [#129](https://github.com/iiasa/ixmp/pull/129), [#132](https://github.com/iiasa/ixmp/pull/132): Fully implement `Scenario.clone()` across platforms (databases).